+++ app/controllers/register/trustees/individual/CheckYourAnswersController.scala
@@ -26,7 +26,7 @@ import models.{CheckMode, Index, NormalMode}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent}
 import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import utils.CheckYourAnswers.Ops._
+import utils.checkyouranswers.Ops._
 import utils.{CountryOptions, Navigator}
 import utils.annotations.TrusteesIndividual
 import viewmodels.{AnswerSection, Message}
