+++ app/identifiers/register/trustees/individual/TrusteePreviousAddressId.scala
@@ -20,7 +20,8 @@ import identifiers._
 import identifiers.register.trustees.TrusteesId
 import models.address.{Address, TolerantAddress}
 import play.api.libs.json.JsPath
-import utils.{CheckYourAnswers, CountryOptions}
+import utils.CountryOptions
+import utils.checkyouranswers.CheckYourAnswers
 
 case class TrusteePreviousAddressId(index: Int) extends TypedIdentifier[Address] {
   override def path: JsPath = TrusteesId(index).path \ TrusteePreviousAddressId.toString
